[leehuun][react-week4-assignment-1]
 - 기존 component refactoring
