+++ src/InputContainer.test.jsx
@@ -1,0 +1,24 @@
+import { fireEvent, render } from '@testing-library/react';
+import { useSelector, useDispatch } from 'react-redux';
+
+import InputContainer from './InputContainer';
+
+jest.mock('react-redux');
+
+test('InputContainer add task', () => {
+  const dispatch = jest.fn();
+  useDispatch.mockImplementation(() => dispatch);
+
+  useSelector.mockImplementation((selector) => selector({ taskTitle: 'New Task' }));
+
+  const { getByText, getByDisplayValue } = render((
+    <InputContainer />
+  ));
+
+  expect(getByText(/추가/)).not.toBeNull();
+  expect(getByDisplayValue(/New Task/)).not.toBeNull();
+
+  fireEvent.click(getByText(/추가/));
+
+  expect(dispatch).toBeCalledWith({ type: 'addTask' });
+});
