+++ src/ListContainer.jsx
@@ -1,0 +1,23 @@
+import { useDispatch, useSelector } from 'react-redux';
+
+import { deleteTask } from './actions';
+import List from './List';
+
+export default function ListContainer() {
+  const { tasks } = useSelector((state) => ({
+    tasks: state.tasks,
+  }));
+
+  const dispatch = useDispatch();
+
+  function handleClickDeleteTask(id) {
+    dispatch(deleteTask(id));
+  }
+
+  return (
+    <List
+      tasks={tasks}
+      onClickDelete={handleClickDeleteTask}
+    />
+  );
+}
