+++ src/ListContainer.test.jsx
@@ -2,11 +2,11 @@ import { render } from '@testing-library/react';
 
 import { useSelector } from 'react-redux';
 
-import App from './App';
+import ListContainer from './ListContainer';
 
 jest.mock('react-redux');
 
-test('App', () => {
+test('ListContainer', () => {
   const tasks = [
     { id: 1, title: '아무 것도 하지 않기 #1' },
     { id: 2, title: '아무 것도 하지 않기 #2' },
@@ -15,13 +15,8 @@ test('App', () => {
   useSelector.mockImplementation((selector) => selector({ tasks }));
 
   const { getByText } = render((
-    <App />
+    <ListContainer />
   ));
 
-  expect(getByText(/추가/)).not.toBeNull();
   expect(getByText(/아무 것도 하지 않기 #1/)).not.toBeNull();
-  expect(getByText(/아무 것도 하지 않기 #2/)).not.toBeNull();
-
-  // TODO: 통합 테스트 코드 작성
-  // CodeceptJS => 실제 브라우저에서 사용자 테스트 실행 가능.
 });
